fix: updated ray to take in float

+++ include/ray.hpp
@@ -8,7 +8,7 @@ struct Ray {
   constexpr Ray(const Point3 &origin, const Vec3 &direction)
       : orig(origin), dir(direction) {}
 
-  constexpr Point3 at(double t) const { return orig + t * dir; }
+  constexpr Point3 at(float t) const { return orig + t * dir; }
 
   const Point3 &origin() const { return orig; }
